@@ -2,6 +2,7 @@
   "name": "livekit-meet",
   "version": "0.2.0",
   "private": true,
+  "type": "module",
   "scripts": {
     "dev": "next dev",
     "build": "next build",
@@ -9,27 +10,77 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@datadog/browser-logs": "^5.23.3",
+    "@datadog/browser-core": "^6.0.0",
+    "@datadog/browser-logs": "^6.0.0",
+    "@hookform/resolvers": "^3.10.0",
     "@livekit/components-react": "2.7.0",
     "@livekit/components-styles": "1.1.4",
     "@livekit/krisp-noise-filter": "0.2.14",
+    "@radix-ui/react-accordion": "^1.2.2",
+    "@radix-ui/react-alert-dialog": "^1.1.4",
+    "@radix-ui/react-aspect-ratio": "^1.1.1",
+    "@radix-ui/react-avatar": "^1.1.2",
+    "@radix-ui/react-checkbox": "^1.1.3",
+    "@radix-ui/react-collapsible": "^1.1.2",
+    "@radix-ui/react-context-menu": "^2.2.4",
+    "@radix-ui/react-dialog": "^1.1.4",
+    "@radix-ui/react-dropdown-menu": "^2.1.4",
+    "@radix-ui/react-hover-card": "^1.1.4",
+    "@radix-ui/react-label": "^2.1.1",
+    "@radix-ui/react-menubar": "^1.1.4",
+    "@radix-ui/react-navigation-menu": "^1.2.3",
+    "@radix-ui/react-popover": "^1.1.4",
+    "@radix-ui/react-progress": "^1.1.1",
+    "@radix-ui/react-radio-group": "^1.2.2",
+    "@radix-ui/react-scroll-area": "^1.2.2",
+    "@radix-ui/react-select": "^2.1.4",
+    "@radix-ui/react-separator": "^1.1.1",
+    "@radix-ui/react-slider": "^1.2.2",
+    "@radix-ui/react-slot": "^1.1.1",
+    "@radix-ui/react-switch": "^1.1.2",
+    "@radix-ui/react-tabs": "^1.1.2",
+    "@radix-ui/react-toast": "^1.2.4",
+    "@radix-ui/react-toggle": "^1.1.1",
+    "@radix-ui/react-toggle-group": "^1.1.1",
+    "@radix-ui/react-tooltip": "^1.1.6",
+    "class-variance-authority": "^0.7.1",
+    "clsx": "^2.1.1",
+    "cmdk": "1.0.0",
+    "embla-carousel-react": "^8.5.2",
+    "immer": "^10.1.1",
+    "input-otp": "^1.4.2",
     "livekit-client": "2.8.0",
     "livekit-server-sdk": "2.9.7",
-    "next": "14.2.23",
-    "react": "18.3.1",
-    "react-dom": "18.3.1",
-    "tinykeys": "^3.0.0"
+    "lucide-react": "^0.473.0",
+    "next": "15.1.4",
+    "next-themes": "^0.4.4",
+    "react": "19.0.0",
+    "react-dom": "19.0.0",
+    "react-hook-form": "^7.54.2",
+    "react-resizable-panels": "^2.1.7",
+    "recharts": "^2.15.0",
+    "sonner": "^1.7.2",
+    "tailwind-merge": "^2.6.0",
+    "tailwindcss-animate": "^1.0.7",
+    "tinykeys": "^3.0.0",
+    "vaul": "^1.1.2",
+    "zod": "^3.24.1",
+    "zustand": "^5.0.3"
   },
   "devDependencies": {
     "@types/node": "22.10.7",
-    "@types/react": "18.3.18",
-    "@types/react-dom": "18.3.5",
+    "@types/react": "19.0.7",
+    "@types/react-dom": "19.0.3",
+    "autoprefixer": "^10.4.20",
     "eslint": "9.18.0",
-    "eslint-config-next": "14.2.23",
+    "eslint-config-next": "15.1.4",
+    "postcss": "^8.5.1",
     "source-map-loader": "^5.0.0",
+    "tailwindcss": "^3.4.17",
     "typescript": "5.7.3"
   },
   "engines": {
     "node": ">=18"
-  }
+  },
+  "packageManager": "pnpm@9.15.4+sha512.b2dc20e2fc72b3e18848459b37359a32064663e5627a51e4c74b2c29dd8e8e0491483c3abb40789cfd578bf362fb6ba8261b05f0387d76792ed6e23ea3b1b6a0"
 }
