@@ -17,7 +17,7 @@ export const useDebugMode = ({ logLevel }: { logLevel?: LogLevel }) => {
       console.log('setting up datadog logs');
       datadogLogs.init({
         clientToken: process.env.NEXT_PUBLIC_DATADOG_CLIENT_TOKEN,
-        site: process.env.NEXT_PUBLIC_DATADOG_SITE,
+        site: process.env.NEXT_PUBLIC_DATADOG_SITE as Parameters<typeof datadogLogs.init>[0]['site'],
         forwardErrorsToLogs: true,
         sessionSampleRate: 100,
       });
